core::packages::PackageInfo
