+++ core/packages/PackageInfo.h
@@ -1,0 +1,15 @@
+#ifndef SORBET_CORE_PACKAGES_PACKAGEINFO_H
+#define SORBET_CORE_PACKAGES_PACKAGEINFO_H
+
+namespace sorbet::core::packages {
+class PackageInfo {
+public:
+    virtual ~PackageInfo() = 0;
+    PackageInfo() = default;
+    PackageInfo(PackageInfo &) = delete;
+    PackageInfo(const PackageInfo &) = delete;
+    PackageInfo &operator=(PackageInfo &&) = delete;
+    PackageInfo &operator=(const PackageInfo &) = delete;
+};
+} // namespace sorbet::core::package
+#endif
